Fix GPU local-dev: shared models, D: cache routing, async inference
- pipeline.py: accept pre-built VAD/ASR/NLLB/TTS instances so all 4
  dialect pipelines share one copy of each model (fits 4 GB VRAM)
- main.py: load shared models once in lifespan; redirect HF_HOME,
  HUGGINGFACE_HUB_CACHE, TORCH_HOME to D: before library imports
- vad.py: torch.hub.set_dir() to D:.cache/torch/hub (C: was full)
- asr.py: pass download_root to whisper.load_model() -> D: cache
- translator.py: replace deprecated torch_dtype= with dtype=
- translate.py: wrap blocking pipeline.run_audio() in run_in_threadpool
  so the async endpoint does not freeze the FastAPI event loop

+++ src/mgvaovao/models/translator.py
@@ -45,7 +45,7 @@ class NLLBTranslator:
 
         # Base model is always NLLB-200; LoRA adapters are applied on top
         base = AutoModelForSeq2SeqLM.from_pretrained(
-            base_name, torch_dtype=dtype, low_cpu_mem_usage=True
+            base_name, dtype=dtype, low_cpu_mem_usage=True
         )
         if ckpt_final.is_dir() and (ckpt_final / "adapter_config.json").is_file():
             from peft import PeftModel
